# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/vip": {
            "get": {
                "description": "获取vip列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vip"
                ],
                "summary": "获取vip列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "获取vip列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vip"
                ],
                "summary": "获取vip列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "vip信息",
                        "name": "Permisson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Permisson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Permisson": {
            "type": "object",
            "properties": {
                "devices": {
                    "description": "设备数",
                    "type": "integer"
                },
                "fence": {
                    "description": "围栏权限 1启用 2禁止",
                    "type": "integer"
                },
                "fenceAlarm": {
                    "description": "围栏报警权限 1启用 2禁止",
                    "type": "integer"
                },
                "id": {
                    "description": "vipID",
                    "type": "integer"
                },
                "logo": {
                    "description": "定制LOGO权限 1启用 2禁止",
                    "type": "integer"
                },
                "orbit": {
                    "description": "轨迹权限 1启用 2禁止",
                    "type": "integer"
                },
                "real": {
                    "description": "实时数据权限 1启用 2禁止",
                    "type": "integer"
                },
                "state": {
                    "description": "定制状态权限 1启用 2禁止",
                    "type": "integer"
                },
                "users": {
                    "description": "子用户数",
                    "type": "integer"
                }
            }
        },
...
    }
}